@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <RollForward>Major</RollForward>
     <PublishReadyToRun>false</PublishReadyToRun>
     <TieredCompilation>false</TieredCompilation>
@@ -19,7 +19,8 @@
     <EmbeddedResource Include="Icon.bmp" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="JitterPhysics" Version="0.2.0.20" />
+    <PackageReference Include="BepuPhysics" Version="2.5.0-beta.23" />
+    <PackageReference Include="BepuUtilities" Version="2.5.0-beta.23" />
     <PackageReference Include="MonoGame.Framework.DesktopGL" Version="3.8.1.303" />
     <PackageReference Include="MonoGame.Content.Builder.Task" Version="3.8.1.303" />
   </ItemGroup>
@@ -30,6 +31,11 @@
   <ItemGroup>
     <Folder Include="Content\Effects\" />
     <Folder Include="Content\Textures\Skybox\" />
+  </ItemGroup>
+  <ItemGroup>
+    <Reference Include="WorldEngine">
+      <HintPath>Dependencies\WorldEngine.dll</HintPath>
+    </Reference>
   </ItemGroup>
   <Target Name="RestoreDotnetTools" BeforeTargets="Restore">
     <Message Text="Restoring dotnet tools" Importance="High" />
