@@ -3,18 +3,20 @@ using Microsoft.Xna.Framework;
 
 namespace TestMonoGame.Game;
 
-public class GameObject : IDisposable
+public class GameObject(string name) : IDisposable
 {
-    public string Id { get; set; }
-    public string Name { get; set; }
+    public string Name = name;
     
-    public readonly Transform Transform = new Transform();
-
-    public virtual void Initialize(Vector3? objectPosition = null, Quaternion? objectRotation = null,
-        Vector3? objectScale = null)
+    public readonly Transform Transform = new()
     {
-        SetPositionAndRotation(objectPosition ?? Vector3.Zero, objectRotation ?? Quaternion.Identity);
-        Transform.Scale = objectScale ?? Vector3.One;
+        Position = Vector3.Zero,
+        Rotation = Quaternion.Identity,
+        Scale = Vector3.One
+    };
+
+    public virtual void Initialize()
+    {
+        
     }
 
     public virtual void Update(GameTime gameTime)
@@ -24,11 +26,5 @@ public class GameObject : IDisposable
     public virtual void Dispose()
     {
         GC.SuppressFinalize(this);
-    }
-
-    public virtual void SetPositionAndRotation(Vector3 position, Quaternion rotation)
-    {
-        Transform.Position = position;
-        Transform.Rotation = rotation;
     }
 }
