@@ -4,19 +4,14 @@ using TestMonoGame.Rendering;
 
 namespace TestMonoGame.Game;
 
-public class MeshObject : GameObject
+public class MeshObject(string name) : GameObject(name)
 {
     public IEffect MeshEffect;
-
-    public Texture2D Texture;
-
-    public Vector2 TextureTiling = new(1f, 1f);
-
-    public Color DiffuseColor = Color.White;
-
-    public bool ReceiveLighting = true;
-
     public Model Model;
+    public Texture2D Texture;
+    public Vector2 TextureTiling = new(1f, 1f);
+    public Color DiffuseColor = Color.White;
+    public bool ReceiveLighting = true;
 
     public virtual void Draw(GraphicsDevice graphicsDevice, GameTime gameTime)
     {
@@ -31,7 +26,8 @@ public class MeshObject : GameObject
         {
             MeshEffect.SetWorldViewProj(Transform.WorldMatrix, Camera.Current.ViewMatrix,
                 Camera.Current.ProjectionMatrix);
-            MeshEffect.SetTexture2D(Texture);
+            if (Texture != null)
+                MeshEffect.SetTexture2D(Texture);
             MeshEffect.SetTiling(TextureTiling.X, TextureTiling.Y);
             MeshEffect.SetDiffuseColor(DiffuseColor);
             MeshEffect.Draw(mesh);
